+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed `ClassCastException` when creating a `Terminal` with the `mordant-jvm-ffm` module on Java 20 and 21. [(#278)](https://github.com/ajalt/mordant/issues/278)
 - Fixed mouse tracking and console mode restoration in raw mode on native Windows targets. [(#293)](https://github.com/ajalt/mordant/issues/293)
 - Fixed native-image builds failing on Windows with GraalVM 25. [(#286)](https://github.com/ajalt/mordant/issues/286)
+- Fixed `readEventOrNull` throwing an exception instead of returning null on timeout on Windows. [(#276)](https://github.com/ajalt/mordant/issues/276)
 
 ## 3.0.2
 ### Changed